requirements.txt
Pillow
selenium
discord
requests